[feat]: Made changes to file names, added examples, moved code snippets and added descriptions to submodules
