+++ examples/pdf_parsing_example.py
@@ -1,0 +1,5 @@
+from src.data_handling.PDFParser import PDFParser
+
+if __name__ == '__main__':
+    pdf_parser = PDFParser()
+    pdf_parser.parse_files('../../../VKRsData')
