--- tests/data_handling/test_persons.json
@@ -1,181 +1,0 @@
-{
-  "123124": {
-    "isu_id": "123124",
-    "data": {
-      "bio": {
-        "jobs": [],
-        "duties": [
-          {
-            "position": {
-              "id": "1452",
-              "name": "член клуба"
-            },
-            "department": {
-              "name": "ITMO.STUDENTS",
-              "id": "30"
-            }
-          },
-          {
-            "position": {
-              "id": "1452",
-              "name": "член клуба"
-            },
-            "department": {
-              "name": "ITMO.STUDENTS",
-              "id": "30"
-            }
-          }
-        ],
-        "education": {
-          "year": "1",
-          "faculty": {
-            "id": "762",
-            "name": "ФПИ и КТ"
-          },
-          "study": "std",
-          "program": {
-            "id": "113633",
-            "name": "Компьютерные технологии в дизайне 2023"
-          }
-        }
-      },
-      "publications": null,
-      "rids": null,
-      "projects": null,
-      "events": null
-    }
-  },
-  "987901": {
-    "isu_id": "987901",
-    "data": {
-      "bio": {
-        "jobs": [],
-        "duties": [],
-        "education": {}
-      },
-      "publications": null,
-      "rids": null,
-      "projects": null,
-      "events": null
-    }
-  },
-  "178321": {
-    "isu_id": "178321",
-    "data": {
-      "bio": {
-        "jobs": [
-          {
-            "position": {
-              "id": 107,
-              "name": "лаборант",
-              "rate": 0.25
-            },
-            "department": {
-              "id": 1161,
-              "name": "лаборатория интеллектуальных сервисов и приложений"
-            }
-          }
-        ],
-        "duties": [],
-        "education": {
-          "year": 1,
-          "faculty": {
-            "id": 725,
-            "name": "ФИКТ"
-          },
-          "study": "std",
-          "program": {
-            "id": 113599,
-            "name": "Мобильные и сетевые технологии 2023"
-          }
-        }
-      },
-      "publications": null,
-      "rids": null,
-      "projects": [
-        {
-          "theme_id": 623090,
-          "type": "НИР",
-          "title": "Методы и алгоритмы интеллектуальных сервисов и приложений",
-          "department_id": 1161,
-          "date_start": "01.09.2023",
-          "date_end": "30.06.2024",
-          "key_words": [
-            "большие языковые модели",
-            "малые объемы доступных данных",
-            "методы и алгоритмы интеллектуальных сервисов и приложений",
-            "рекомендательные системы",
-            "робастность"
-          ],
-          "role": "Участник",
-          "customer": "Университет ИТМО"
-        }
-      ],
-      "events": [
-        {
-          "title": "XIII Конгресс молодых ученых ИТМО",
-          "year": 2024,
-          "type": "Конгресс",
-          "rank": "всероссийский",
-          "role": "Спикер",
-          "date_start": "08 Апреля",
-          "date_end": "11 Апреля"
-        },
-        {
-          "title": "ITMO.Megabattle Раунд 1 (2023)",
-          "year": 2023,
-          "type": "Фестиваль",
-          "rank": "университетский",
-          "role": "Участник",
-          "date_start": "09 Сентября",
-          "date_end": "30 Ноября"
-        },
-        {
-          "title": "Первенство Санкт-Петербурга по спортивному ориентированию  (кросс-классика - общий старт)",
-          "year": 2023,
-          "type": "Соревнование",
-          "rank": "городской",
-          "role": "Участник",
-          "date_start": "01 Октября",
-          "date_end": "01 Октября"
-        },
-        {
-          "title": "Чемпионат Университета ИТМО по киберспорту (Kronbars Games) (1 круг) (CS:GO)",
-          "year": 2023,
-          "type": "Соревнование",
-          "rank": "университетский",
-          "role": "Участник",
-          "date_start": "17 Октября",
-          "date_end": "24 Декабря"
-        },
-        {
-          "title": "Чемпионат Университета ИТМО по киберспорту (Kronbars Games) (1 круг) (Among Us)",
-          "year": 2023,
-          "type": "Соревнование",
-          "rank": "университетский",
-          "role": "Участник",
-          "date_start": "17 Октября",
-          "date_end": "24 Декабря"
-        },
-        {
-          "title": "Открытый Чемпионат Университета ИТМО по шахматам в рамках Лиги Нового ФизТеха (1 круг) (Парные шахматы)",
-          "year": 2023,
-          "type": "Соревнование",
-          "rank": "городской",
-          "role": "Участник",
-          "date_start": "08 Октября",
-          "date_end": "17 Декабря"
-        },
-        {
-          "title": "Международная олимпиада Изумруд по информатике 2023",
-          "year": 2023,
-          "type": "Олимпиада",
-          "rank": "международный",
-          "role": "Участник",
-          "date_start": "01 Февраля",
-          "date_end": "10 Апреля"
-        }
-      ]
-    }
-  }
-}
